MERGE SORT
Merge Sort is a highly efficient, general-purpose, comparison-based sorting algorithm. It is one of the most classic examples of the Divide and Conquer paradigm.It's a stable sort (maintaining the relative order of equal elements) and is generally not an in-place algorithm, as it requires temporary storage space proportional to the input size ($O(n)$) to merge the sorted sub-arrays.