BINARY SEARCH TREE
BST traversal refers to the process of visiting (or examining) every node in the tree exactly once. Unlike linear data structures, which have a single, sequential way to traverse them, trees can be traversed in multiple ways. The three most common ways to traverse a BST are In-order, Pre-order, and Post-order traversal, all of which use a Depth-First Search (DFS) approach.